various small things
