+++ src/hepattn/experiments/trackml/eval/plot_event.py
