+++ src/hepattn/experiments/trackml/eval/plot_kinematics.py
