+++ src/hepattn/experiments/trackml/submit/submit_prep_rcif.sh
